// src/utils/customer/types.ts
export type CreateCustomerParams = {
    
    user_id: number

    customer_name: string

    customer_contact: string

    customer_email: string

    username: string

    password: string

    account_status: number
}

export type UpdateCustomerParams = {
    
    user_id: number

    customer_name: string

    customer_contact: string

    customer_email: string

    username: string

    password: string

    account_status: number
}